+++ imblearn/utils/__init__.py
@@ -1,7 +1,0 @@
-"""
-The :mod:`imblearn.utils` module includes various utilities.
-"""
-
-from .validation import check_target_type
-
-__all__ = ['check_target_type']